refactor: split snappable
